refactor: pulumi setup rearranged
