+++ infra/vf-mvp/resources/fargate.ts
@@ -1,0 +1,105 @@
+import * as aws from '@pulumi/aws';
+import * as awsx from '@pulumi/awsx';
+import * as pulumi from '@pulumi/pulumi';
+import setup, { nameResource } from '../utils/setup';
+
+const {
+  tags,
+  envOverride,
+  externalApis: { codesetsEndpoint, usersApiEndpoint },
+  backendSignKey
+} = setup;
+
+export function createFargateService(loadBalancer: awsx.lb.ApplicationLoadBalancer, cdn: aws.cloudfront.Distribution) {
+
+  // ECS Task role
+  const awsIdentity = pulumi.output(aws.getCallerIdentity());
+  const taskRole = new aws.iam.Role(nameResource('fargate-task-role'), {
+    tags,
+    assumeRolePolicy: aws.iam.assumeRolePolicyForPrincipal({
+      Service: 'ecs-tasks.amazonaws.com',
+    }),
+  });
+
+  const sinunaAccessPolicy = new aws.iam.Policy(
+    nameResource('fargate-task-role-sinuna-policy'),
+    {
+      tags,
+      policy: {
+        Version: '2012-10-17',
+        Statement: [
+          {
+            Effect: 'Allow',
+            Action: ['ssm:GetParameter'],
+            Resource: [
+              pulumi.interpolate`arn:aws:ssm:${aws.config.region}:${awsIdentity.accountId}:parameter/${envOverride}_SINUNA_CLIENT_ID`, // Access to stage-prefixed sinuna variables
+              pulumi.interpolate`arn:aws:ssm:${aws.config.region}:${awsIdentity.accountId}:parameter/${envOverride}_SINUNA_CLIENT_SECRET`,
+            ],
+          },
+        ],
+      },
+    }
+  );
+
+  // Attach a policy to grant access to Parameter Store
+  new aws.iam.RolePolicyAttachment(nameResource('fargate-task-role-sinuna-policy-attachment'), {
+    role: taskRole.name,
+    policyArn: sinunaAccessPolicy.arn,
+  });
+
+  const testbedConfig = new pulumi.Config("testbed");
+
+  // ECR repository
+  const repository = new awsx.ecr.Repository(nameResource('ecr-repo'), {
+    tags,
+    forceDelete: true,
+  });
+
+  // ECR Docker image
+  const image = new awsx.ecr.Image(nameResource('mvp-image'), {
+    repositoryUrl: repository.url,
+    path: '../../', // path to a directory to use for the Docker build context (root of the repo)
+    dockerfile: '../../apps/vf-mvp/Dockerfile', // dockerfile may be used to override the default Dockerfile name and/or location
+    extraOptions: ['--platform', 'linux/amd64'],
+    args: {
+      NEXT_PUBLIC_CODESETS_BASE_URL: codesetsEndpoint,
+      NEXT_PUBLIC_USERS_API_BASE_URL: usersApiEndpoint,
+      BACKEND_SECRET_SIGN_KEY: backendSignKey,
+      NEXT_PUBLIC_STAGE: envOverride,
+      TESTBED_PRODUCT_GATEWAY_BASE_URL: process.env.TESTBED_PRODUCT_GATEWAY_BASE_URL || testbedConfig.require("gatewayUrl"),
+      TESTBED_DEFAULT_DATA_SOURCE: process.env.TESTBED_DEFAULT_DATA_SOURCE || testbedConfig.require("defaultDataSource"),
+      FRONTEND_ORIGIN_URI: pulumi.interpolate`https://${cdn.domainName}`,
+    },
+  });
+
+  // ECS cluster
+  const cluster = new aws.ecs.Cluster(nameResource('ecs-cluster'), {
+    tags,
+  });
+
+  // Fargate service
+  return new awsx.ecs.FargateService(
+    nameResource('fargate-service'),
+    {
+      tags,
+      cluster: cluster.arn,
+      assignPublicIp: true,
+      continueBeforeSteadyState: false,
+      taskDefinitionArgs: {
+        containers: {
+          service: {
+            image: image.imageUri,
+            portMappings: [
+              {
+                targetGroup: loadBalancer.defaultTargetGroup,
+              },
+            ],
+          },
+        },
+        taskRole: {
+          roleArn: taskRole.arn,
+        },
+      },
+    }
+  );
+}
